allow custom YamlConvert function for conversions
in Utilities.TypeConverter.ChangeType this allows additionally defining a YamlConvert method instead of overriding the implicit or explicit cast operators. Reason for this is that the implicit/explicit operators don't allow casting from interfaces or object.

Reason for using this instead of the type converters is types like some generic `Either<A, B>` class which can deserialize from multiple different input types, which isn't achievable using type converters.

This YamlConvert method can be defined both on the source or the target type.